feat: adiciona página de gerenciamento de tags com componentes de criação e listagem
