+++ src/pages/admin/tags/PageTags.tsx
@@ -1,0 +1,14 @@
+import { LayoutDashboard } from '../../../shared/layouts';
+import { CreateTag } from './components/CreateTag';
+import { ListTags } from './components/ListTags';
+
+export const PageTags = () => {
+  return (
+    <LayoutDashboard>
+      <div className="container mx-auto px-4 grid gap-6 md:grid-cols-2">
+        <CreateTag />
+        <ListTags />
+      </div>
+    </LayoutDashboard>
+  );
+};
