+++ src/routes/index.tsx
@@ -12,7 +12,7 @@ import { PublishedPills } from '../pages/posts/PublishedPills';
 import { PillDetails } from '../pages/posts/PillDetails';
 import { PostDetails } from '../pages/posts/PostDetails';
 import { PostEditor } from '../pages/dashboard/PostEditor';
-import { ListagemTags } from '../pages/dashboard/ListagemTags';
+import { PageTags } from '../pages/admin/tags/PageTags';
 
 type TPrivateRouteProps = RouteProps & {
   element: React.ReactNode;
@@ -51,7 +51,7 @@ export const AppRoutes = () => {
 
       {/* Login Admin */}
       <Route path="/admin/login" element={<AdminLogin />} />
-      <Route path="/admin/tags" element={<PrivateRoute element={<ListagemTags />} />} />
+      <Route path="/admin/tags" element={<PrivateRoute element={<PageTags />} />} />
 
       {/* Página em Branco (Para teste ou outro propósito) */}
       <Route path="/blank" element={<BlankPage />} />
